requirements.txt
absl-py==0.2.2
albumentations==0.2.3
amqp==2.3.2
asn1crypto==0.24.0
astor==0.6.2
backcall==0.1.0
billiard==3.5.0.4
bleach==1.5.0
blinker==1.3
celery==4.2.0
certifi==2018.8.13
cffi==1.12.1
chardet==3.0.4
cinbase==1.0.2
click==6.7
cloud-init==18.4
cloudpickle==0.5.3
command-not-found==0.3
ConfigArgParse==0.14.0
configobj==5.0.6
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
cycler==0.10.0
Cython==0.28.4
dask==0.18.2
dataclasses==0.6
decorator==4.3.0
defusedxml==0.5.0
Distance==0.1.3
Django==1.11.20
django-ckeditor==5.3.0
django-crum==0.7.2
django-jet==1.0.6
django-js-asset==0.1.1
django-rest-framework==0.1.0
django-tinymce==2.6.0
django-user-agents==0.3.2
djangorestframework==3.6.3
djongo==1.2.26
dominate==2.3.5
elasticsearch==6.2.0
entrypoints==0.3
enum-compat==0.0.2
et-xmlfile==1.0.1
eventlet==0.21.0
Flask==1.0.2
future==0.16.0
fuzzywuzzy==0.17.0
gast==0.2.0
greenlet==0.4.15
grpcio==1.12.1
h5py==2.7.1
html5lib==0.9999999
idna==2.7
imgaug==0.2.6
imutils==0.4.6
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
itypes==1.1.0
jaconv==0.2.3
jdcal==1.4
jedi==0.13.2
Jinja2==2.10.1
json2html==1.2.1
jsonpatch==1.10
jsonpointer==1.9
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
keras-tcn==2.3.4
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
kombu==4.2.0
language-selector==0.1
lib-ocr==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
meld3==1.0.2
mistune==0.8.4
mpmath==1.1.0
mysql-connector-python==8.0.13
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.1
notebook==5.7.4
numpy==1.15.0
oauthlib==1.0.3
olefile==0.46
opencv-python==3.4.2.17
opencv-python-headless==3.4.2.17
OpenNMT-py==0.6.0
openpyxl==2.4.9
pandas==0.21.1
pandocfilters==1.4.2
parso==0.3.4
pdfminer.six==20181108
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prettytable==0.7.2
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.0
psutil==5.5.1
ptyprocess==0.6.0
pyasn1==0.1.9
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.0
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.20.0
PyICU==2.0.3
PyJWT==1.3.0
pylsd==0.0.2
pymongo==3.7.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyserial==3.0.1
pytesseract==0.2.6
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.7.3
python-debian==0.1.27
python-igraph==0.7.1.post6
python-Levenshtein==0.12.0
python-systemd==231
pytz==2018.5
PyWavelets==0.5.2
pyxdg==0.25
PyYAML==4.2b1
pyzmq==17.1.2
qtconsole==4.4.3
redis==2.10.6
regex==2018.2.21
requests==2.20.0
rstr==2.2.6
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8.1
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.11.0
sklearn==0.0
sortedcontainers==2.0.5
sqlparse==0.2.4
ssh-import-id==5.5
sympy==1.3
tensorboard==1.8.0
tensorflow==1.12.2
tensorflow-gpu==1.12.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textdistance==3.1.0
Theano==1.0.2
toolz==0.9.0
torch==1.0.1.post2
torchtext==0.3.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.24.0
traitlets==4.3.2
ua-parser==0.8.0
ufw==0.35
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.2
user-agents==1.1.0
uWSGI==2.0.15
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1
widgetsnbextension==3.4.2
xlrd==1.1.0
XlsxWriter==1.0.5
